pyproject.toml
[project]
name = "agents-playground"
version = "0.1.0"
description = "Local notebook experiments for ai-chat"
requires-python = ">=3.12,<3.13"
dependencies = [
    "datasets>=4.8.4",
    "openai>=2.0.0",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.12.0",
]

[tool.uv]
package = false
